@@ -26,12 +26,12 @@
     "karma-jasmine": "^0.3.6",
     "karma-phantomjs-launcher": "^0.2.1",
     "phantomjs": "^1.9.18",
-    "react": "^0.14.0",
+    "react": "^15.3.2",
     "react-component-gulp-tasks": "^0.7.6",
-    "react-dom": "^0.14.0"
+    "react-dom": "^15.3.2"
   },
   "peerDependencies": {
-    "react": "^0.14.0"
+    "react": "^15.3.2"
   },
   "browserify-shim": {
     "react": "global:React"
