@@ -6,9 +6,17 @@ Object.defineProperty(exports, '__esModule', {
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
+var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+
+var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var _react = require('react');
 
@@ -18,241 +26,272 @@ var _reactDom = require("react-dom");
 
 var _reactDom2 = _interopRequireDefault(_reactDom);
 
-var _dateHeaderJsx = require('./date-header.jsx');
+var _dateHeaderJs = require('./date-header.js');
 
-var _dateHeaderJsx2 = _interopRequireDefault(_dateHeaderJsx);
+var _dateHeaderJs2 = _interopRequireDefault(_dateHeaderJs);
 
-var _dateCalendarJsx = require('./date-calendar.jsx');
+var _dateCalendarJs = require('./date-calendar.js');
 
-var _dateCalendarJsx2 = _interopRequireDefault(_dateCalendarJsx);
+var _dateCalendarJs2 = _interopRequireDefault(_dateCalendarJs);
 
-var _dateInputJsx = require('./date-input.jsx');
+var _dateInputJs = require('./date-input.js');
 
-var _dateInputJsx2 = _interopRequireDefault(_dateInputJsx);
+var _dateInputJs2 = _interopRequireDefault(_dateInputJs);
 
-var _dateInBodyJsx = require('./date-in-body.jsx');
+var _dateInBodyJs = require('./date-in-body.js');
 
-var _dateInBodyJsx2 = _interopRequireDefault(_dateInBodyJsx);
+var _dateInBodyJs2 = _interopRequireDefault(_dateInBodyJs);
 
 var now = new Date();
-var ReactDatepickerPlus = _react2['default'].createClass({
-	displayName: 'ReactDatepickerPlus',
 
-	propTypes: {
-		selected: _react2['default'].PropTypes.object, //default date
-		format: _react2['default'].PropTypes.string, //format date
-		isfill: _react2['default'].PropTypes.bool, //show prefix-prev prefix-next month
-		months: _react2['default'].PropTypes.number, //show multi-panes by months
+var ReactDatepickerPlus = (function (_Component) {
+	_inherits(ReactDatepickerPlus, _Component);
 
-		time: _react2['default'].PropTypes.bool, //show time select @todo
+	// propTypes: {
+	// 	selected: React.PropTypes.object,	//default date
+	// 	format: React.PropTypes.string,     //format date
+	// 	isfill: React.PropTypes.bool,	    //show prefix-prev prefix-next month
+	// 	months: React.PropTypes.number,		//show multi-panes by months
 
-		min: _react2['default'].PropTypes.object, //select date range min
-		max: _react2['default'].PropTypes.object, //select date range max
+	// 	time: React.PropTypes.bool,         //show time select @todo
 
-		start: _react2['default'].PropTypes.object, //selected time is a range, start date
-		end: _react2['default'].PropTypes.object, //selected time is a range, start date
+	// 	min: React.PropTypes.object,        //select date range min
+	// 	max: React.PropTypes.object,        //select date range max
 
-		className: _react2['default'].PropTypes.string, // custom class
-		disabled: _react2['default'].PropTypes.bool, //input can't change
-		autoHide: _react2['default'].PropTypes.bool, //selected auto hide
-		inline: _react2['default'].PropTypes.bool, //inline
-		lang: _react2['default'].PropTypes.oneOf(['cn', 'en']),
+	// 	start: React.PropTypes.object,	    //selected time is a range, start date
+	// 	end: React.PropTypes.object,	    //selected time is a range, start date
 
-		festival: _react2['default'].PropTypes.bool, //show festival
-		haslunar: _react2['default'].PropTypes.bool, //show lunar
+	// 	className: React.PropTypes.string,  // custom class
+	// 	disabled: React.PropTypes.bool,     //input can't change
+	// 	autoHide: React.PropTypes.bool,     //selected auto hide
+	// 	inline: React.PropTypes.bool,       //inline
+	// 	lang: React.PropTypes.oneOf(['cn', 'en']),
 
-		onFocus: _react2['default'].PropTypes.func, //args (event, picker)
-		onBlur: _react2['default'].PropTypes.func, //args (event, picker)
-		onChange: _react2['default'].PropTypes.func, //args (dayinfo, picker)
-		dayAddon: _react2['default'].PropTypes.func //args (dayinfo)
-		//dayAddon, add data for day, and need to return dom,
-		//the return value will be insert to day each element. pls see last demo
-	},
-	getInitialState: function getInitialState() {
-		return {
+	// 	festival: React.PropTypes.bool, 	//show festival
+	// 	haslunar: React.PropTypes.bool, 	//show lunar
+
+	// 	onFocus: React.PropTypes.func,		//args (event, picker)
+	// 	onBlur: React.PropTypes.func,		//args (event, picker)
+	// 	onChange: React.PropTypes.func,		//args (dayinfo, picker)
+	// 	dayAddon: React.PropTypes.func 		//args (dayinfo)
+	// 	//dayAddon, add data for day, and need to return dom,
+	// 	//the return value will be insert to day each element. pls see last demo
+	// }
+
+	function ReactDatepickerPlus(props) {
+		_classCallCheck(this, ReactDatepickerPlus);
+
+		_get(Object.getPrototypeOf(ReactDatepickerPlus.prototype), 'constructor', this).call(this, props);
+		this.state = {
 			date: now, //render month by date
-			show: this.props.inline ? true : false,
+			show: props.inline ? true : false,
 			focus: false, //focus state
 			offset: {}, //datepicker position
-			selected: this.props.selected,
-			start: this.props.start,
-			end: this.props.end,
-			status: '' };
-	},
-
-	//React.PropTypes.oneOf(['start', 'end', ''])
-	//'start' and 'end' use by bi-datepicker range value, and '' use by single datepicker single date
-	getDefaultProps: function getDefaultProps() {
-		return {
-			isfill: false,
-			format: 'yyyy-MM-dd',
-			selected: now,
-			months: 1,
-			lang: 'en'
+			selected: props.selected,
+			start: props.start,
+			end: props.end,
+			status: '' //React.PropTypes.oneOf(['start', 'end'])
+			//'start' and 'end' use by bi-datepicker range value, and undefined use by single datepicker single date
 		};
-	},
-	onFocus: function onFocus(event, input) {
-		var _state = this.state;
-		var show = _state.show;
-		var focus = _state.focus;
-		var selected = _state.selected;
-
-		if (show && !focus) {
-			this.state.focus = true; //just change state not trigger render
-			return;
-		}
-		var status = input.props.status;selected = status ? this.state[status] : selected;
-		var size = input.handlePosition();
-		var left = size.left;
-		var top = size.top;
-		var height = size.height;
-		top += height + (document.body.scrollTop || document.documentElement.scrollTop);
-
-		var onFocus = this.props.onFocus;
-
-		this.show(true, { left: left, top: top }, true, status);
-		onFocus && onFocus(event, this);
-		if (status) this.setState({ date: selected });
-	},
-	onBlur: function onBlur(event, input) {
-		var _state2 = this.state;
-		var show = _state2.show;
-		var focus = _state2.focus;
-		var _props = this.props;
-		var inline = _props.inline;
-		var onBlur = _props.onBlur;
-
-		if (!show) return;
-		if (!focus) {
-			input.focus(); //when show && !focus, trigger focus 
-		} else if (!inline) {
-				onBlur && onBlur(event, this);
-				focus && this.removePicker();
-			}
-	},
-	show: function show(_show, offset, focus, status) {
-		this.setState({ show: _show, offset: offset, focus: focus, status: status });
-	},
-	updateMonth: function updateMonth(num) {
-		var date = this.state.date;
-
-		var cdate = this.numMonth(date, num);
-		this.updateDate({ date: cdate }, true);
-	},
-	numMonth: function numMonth(date, num) {
-		return new Date(date.getFullYear(), date.getMonth() + num, date.getDate());
-	},
-	updateDay: function updateDay(dateinfo) {
-		this.updateDate(dateinfo);
-	},
-	updateDate: function updateDate(dateinfo, isMonth) {
-		var _props2 = this.props;
-		var onChange = _props2.onChange;
-		var autoHide = _props2.autoHide;
-		var _state3 = this.state;
-		var status = _state3.status;
-		var selected = _state3.selected;
-
-		var getSelected = !isMonth ? dateinfo.date : this.state[status ? status : 'selected'];
-		// let temp = {}; temp[status] = getSelected
-		this.setState(_defineProperty({ show: true, date: dateinfo.date, selected: getSelected, focus: false }, status, getSelected));
-		onChange && onChange(dateinfo, this);
-		!isMonth && autoHide && this.removePicker();
-	},
-	removePicker: function removePicker() {
-		this.show(false);
-		this.refs.insDateInBody && this.refs.insDateInBody.removePicker(true);
-	},
-
-	pickers: function pickers(status) {
-		var $pickers = [],
-		    offsets = [],
-		    dh = undefined,
-		    dc = undefined,
-		    idate = undefined;
-		var _state4 = this.state;
-		var date = _state4.date;
-		var start = _state4.start;
-		var end = _state4.end;
-		var offset = _state4.offset;
-		var _props3 = this.props;
-		var inline = _props3.inline;
-		var months = _props3.months;
-		var lang = _props3.lang;
-		var haslunar = _props3.haslunar;
-		var className = _props3.className;
-
-		var selected = this.state[status ? status : 'selected'];
-		var classes = 'date-picker date-picker-' + (inline ? 'inline' : 'block') + ' ' + (className ? className : '') + ' ' + (haslunar ? 'date-picker-lunar' : '');
-		for (var i = 0; i < months; i++) {
-			offsets.push({ left: i * 215 + offset.left, top: offset.top });
-			idate = this.numMonth(date, i);
-			dh = _react2['default'].createElement(_dateHeaderJsx2['default'], { date: idate, lang: haslunar ? 'cn' : lang, updateMonth: this.updateMonth });
-			dc = _react2['default'].createElement(_dateCalendarJsx2['default'], _extends({}, this.props, { date: idate, status: status, start: start, end: end, selected: selected, onChange: this.updateDay }));
-
-			$pickers.push(inline ? _react2['default'].createElement(
-				'div',
-				{ className: classes, key: i },
-				dh,
-				dc
-			) : _react2['default'].createElement(
-				'div',
-				{ className: classes, style: offsets[i], key: i },
-				dh,
-				dc
-			));
-		}
-		return $pickers;
-	},
-	render: function render() {
-		var _this = this;
-
-		var _state5 = this.state;
-		var show = _state5.show;
-		var selected = _state5.selected;
-		var start = _state5.start;
-		var end = _state5.end;
-		var status = _state5.status;
-		var _props4 = this.props;
-		var format = _props4.format;
-		var inline = _props4.inline;
-		var months = _props4.months;
-		var disabled = _props4.disabled;
-
-		var picker = undefined,
-		    pickers = undefined,
-		    pickerInBody = undefined;
-		var di = function di(val, stat) {
-			return _react2['default'].createElement(_dateInputJsx2['default'], { selected: !val ? selected : val, format: format, disabled: disabled,
-				onFocus: _this.onFocus, onBlur: _this.onBlur, status: stat });
-		};
-		if (show) {
-			pickers = this.pickers(status);
-			picker = _react2['default'].createElement(
-				'div',
-				{ className: months > 1 ? 'date-multi clearfix' : '' },
-				pickers
-			);
-			pickerInBody = _react2['default'].createElement(
-				_dateInBodyJsx2['default'],
-				{ classes: 'date-picker-wrapper', ref: 'insDateInBody' },
-				picker
-			);
-		}
-		var didom = start && end ? _react2['default'].createElement(
-			'div',
-			{ className: 'date-inputs' },
-			di(start, 'start'),
-			di(end, 'end')
-		) : di();
-		return _react2['default'].createElement(
-			'div',
-			{ className: 'date-components' },
-			didom,
-			inline ? picker : pickerInBody
-		);
 	}
-});
+
+	_createClass(ReactDatepickerPlus, [{
+		key: 'onFocus',
+		value: function onFocus(event, input) {
+			var _state = this.state;
+			var show = _state.show;
+			var focus = _state.focus;
+			var selected = _state.selected;
+
+			if (show && !focus) {
+				this.state.focus = true; //just change state not trigger render
+				return;
+			}
+			var status = input.props.status;selected = status ? this.state[status] : selected;
+
+			var _input$handlePosition = input.handlePosition();
+
+			var left = _input$handlePosition.left;
+			var top = _input$handlePosition.top;
+			var height = _input$handlePosition.height;
+
+			top += height + (document.body.scrollTop || document.documentElement.scrollTop);
+			var onFocus = this.props.onFocus;
+
+			this.show(true, { left: left, top: top }, true, status);
+			onFocus && onFocus(event, this);
+			if (status) this.setState({ date: selected });
+		}
+	}, {
+		key: 'onBlur',
+		value: function onBlur(event, input) {
+			var _state2 = this.state;
+			var show = _state2.show;
+			var focus = _state2.focus;
+			var _props = this.props;
+			var inline = _props.inline;
+			var onBlur = _props.onBlur;
+
+			if (!show) return;
+			if (!focus) {
+				input.focus(); //when show && !focus, trigger focus 
+			} else if (!inline) {
+					onBlur && onBlur(event, this);
+					focus && this.removePicker();
+				}
+		}
+	}, {
+		key: 'show',
+		value: function show(_show, offset, focus, status) {
+			this.setState({ show: _show, offset: offset, focus: focus, status: status });
+		}
+	}, {
+		key: 'updateMonth',
+		value: function updateMonth(num) {
+			var date = this.state.date;
+
+			var cdate = this.numMonth(date, num);
+			this.updateDate({ date: cdate }, true);
+		}
+	}, {
+		key: 'numMonth',
+		value: function numMonth(date, num) {
+			return new Date(date.getFullYear(), date.getMonth() + num, date.getDate());
+		}
+	}, {
+		key: 'updateDay',
+		value: function updateDay(dateinfo) {
+			this.updateDate(dateinfo);
+		}
+	}, {
+		key: 'updateDate',
+		value: function updateDate(dateinfo, isMonth) {
+			var _props2 = this.props;
+			var onChange = _props2.onChange;
+			var autoHide = _props2.autoHide;
+			var _state3 = this.state;
+			var _state3$status = _state3.status;
+			var status = _state3$status === undefined ? 'selected' : _state3$status;
+			var selected = _state3.selected;
+
+			var getSelected = !isMonth ? dateinfo.date : this.state[status];
+			// let temp = {}; temp[status] = getSelected
+			this.setState(_defineProperty({ show: true, date: dateinfo.date, selected: getSelected, focus: false }, status, getSelected));
+			onChange && onChange(dateinfo, this);
+			!isMonth && autoHide && this.removePicker();
+		}
+	}, {
+		key: 'removePicker',
+		value: function removePicker() {
+			this.show(false);
+			this.refs.insDateInBody && this.refs.insDateInBody.removePicker(true);
+		}
+	}, {
+		key: 'pickers',
+		value: function pickers(status) {
+			var $pickers = [],
+			    offsets = [],
+			    dh = undefined,
+			    dc = undefined,
+			    idate = undefined;
+			var _state4 = this.state;
+			var date = _state4.date;
+			var start = _state4.start;
+			var end = _state4.end;
+			var offset = _state4.offset;
+			var _props3 = this.props;
+			var inline = _props3.inline;
+			var months = _props3.months;
+			var lang = _props3.lang;
+			var haslunar = _props3.haslunar;
+			var className = _props3.className;
+
+			var selected = this.state[status ? status : 'selected'];
+			var classes = 'date-picker date-picker-' + (inline ? 'inline' : 'block') + ' ' + (className ? className : '') + ' ' + (haslunar ? 'date-picker-lunar' : '');
+			for (var i = 0; i < months; i++) {
+				offsets.push({ left: i * 215 + offset.left, top: offset.top });
+				idate = this.numMonth(date, i);
+				dh = _react2['default'].createElement(_dateHeaderJs2['default'], { date: idate, lang: haslunar ? 'cn' : lang, updateMonth: this.updateMonth });
+				dc = _react2['default'].createElement(_dateCalendarJs2['default'], _extends({}, this.props, { date: idate, status: status, start: start, end: end, selected: selected, onChange: this.updateDay.bind(this) }));
+
+				$pickers.push(inline ? _react2['default'].createElement(
+					'div',
+					{ className: classes, key: i },
+					dh,
+					dc
+				) : _react2['default'].createElement(
+					'div',
+					{ className: classes, style: offsets[i], key: i },
+					dh,
+					dc
+				));
+			}
+			return $pickers;
+		}
+	}, {
+		key: 'render',
+		value: function render() {
+			var _this = this;
+
+			var _state5 = this.state;
+			var show = _state5.show;
+			var selected = _state5.selected;
+			var start = _state5.start;
+			var end = _state5.end;
+			var status = _state5.status;
+			var _props4 = this.props;
+			var format = _props4.format;
+			var inline = _props4.inline;
+			var months = _props4.months;
+			var disabled = _props4.disabled;
+
+			var picker = undefined,
+			    pickers = undefined,
+			    pickerInBody = undefined;
+			var di = function di(val, stat) {
+				return _react2['default'].createElement(_dateInputJs2['default'], { selected: !val ? selected : val, format: format, disabled: disabled,
+					onFocus: _this.onFocus.bind(_this), onBlur: _this.onBlur.bind(_this), status: stat });
+			};
+			if (show) {
+				pickers = this.pickers(status);
+				picker = _react2['default'].createElement(
+					'div',
+					{ className: months > 1 ? 'date-multi clearfix' : '' },
+					pickers
+				);
+				pickerInBody = _react2['default'].createElement(
+					_dateInBodyJs2['default'],
+					{ className: 'date-picker-wrapper', ref: 'insDateInBody' },
+					picker
+				);
+			}
+			var didom = start && end ? _react2['default'].createElement(
+				'div',
+				{ className: 'date-inputs' },
+				di(start, 'start'),
+				di(end, 'end')
+			) : di();
+			return _react2['default'].createElement(
+				'div',
+				{ className: 'date-components' },
+				didom,
+				inline ? picker : pickerInBody
+			);
+		}
+	}]);
+
+	return ReactDatepickerPlus;
+})(_react.Component);
+
+;
+
+ReactDatepickerPlus.defaultProps = {
+	isfill: false,
+	format: 'yyyy-MM-dd',
+	selected: now,
+	months: 1,
+	lang: 'en'
+};
 
 exports['default'] = ReactDatepickerPlus;
 module.exports = exports['default'];
