@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { Component } from 'react'
 const chars = {
 	weeksCn:['日', '一', '二', '三', '四', '五', '六'],
     weeksEnF: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
@@ -6,16 +6,16 @@ const chars = {
     monthsCn:['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],
     monthsEn: ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December']
 }
-var dateHeader = React.createClass({
-	propTypes: {
-	    updateMonth: React.PropTypes.func
-	},
+class dateHeader extends Component {
+	// propTypes: {
+	//     updateMonth: React.PropTypes.func
+	// },
 	getDate(){
 		return this.props.date
-	},
+	}
 	changeMonth(num){
 		this.props.updateMonth(num)
-	},
+	}
 	render () {
 		const {lang} = this.props
 		const cn = lang === 'cn'
@@ -33,6 +33,6 @@ var dateHeader = React.createClass({
 			</div>
 		</div>
 	}
-});
+};
 
 export default dateHeader
